requirements.txt
openai
pytesseract
Pillow
pdfplumber
pandas
sqlalchemy
python-dotenv
fastapi
uvicorn
openpyxl